chore(refactor): refactor basecode
